+++ changelog.md
@@ -2,6 +2,10 @@
 
 A list of what's changed, and more.
 
+## 1.1.7
+
+- Preserves inline echos as text (#25)
+
 ## 1.1.6
 
 - Improves wrapping behavior inside conditions (#26)